requirements.txt
rich>=13.7.0
pyyaml>=6.0.1
pyodbc>=5.1.0
colorama>=0.4.6

